move components into component directories
