+++ examples/css-keyframes/src/components/Canvas/Canvas.styles.ts
