+++ examples/css-keyframes/src/components/Canvas/Canvas.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { TimelineStateContext } from "timelime/react";
-import s from "~examples/css-keyframes/Canvas.styles";
+import s from "~examples/css-keyframes/components/Canvas/Canvas.styles";
 
 export const Canvas: React.FC = () => {
   const { GraphEditor } = useContext(TimelineStateContext)!;
