+++ examples/css-keyframes/src/components/Layout/Layout.styles.ts
