+++ examples/css-keyframes/src/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import s from "~examples/css-keyframes/Layout/Layout.styles";
+import s from "~examples/css-keyframes/components/Layout/Layout.styles";
 
 export const Layout: React.FC = (props) => {
   return <div className={s("container")}>{props.children}</div>;
