+++ examples/css-keyframes/src/components/NumberInput/NumberInput.styles.ts
