+++ examples/css-keyframes/src/components/Preview/Preview.styles.ts
