+++ examples/css-keyframes/src/components/Property/Property.styles.ts
