+++ examples/css-keyframes/src/components/PropertyList/PropertyList.styles.ts
