+++ examples/css-keyframes/src/components/Timeline/Timeline.styles.ts
