+++ examples/css-keyframes/src/components/TimelineProperties/TimelineProperties.styles.ts
