+++ examples/css-keyframes/src/components/TimelineProperties/TimelineProperties.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useTimelineLength } from "timelime/react";
-import { NumberInput } from "~examples/css-keyframes/NumberInput/NumberInput";
-import s from "~examples/css-keyframes/TimelineProperties.styles";
+import { NumberInput } from "~examples/css-keyframes/components/NumberInput/NumberInput";
+import s from "~examples/css-keyframes/components/TimelineProperties/TimelineProperties.styles";
 
 export const TimelineProperties: React.FC = () => {
   const [length, setLength] = useTimelineLength();
